+++ dev/clj/user.clj
@@ -6,7 +6,8 @@
     [clojure.spec :as s]
     [clojure.spec.gen :as sg]
     [clojure.tools.namespace.repl :refer [refresh]]
-    [figwheel-sidecar.repl-api :as ra]))
+    [figwheel-sidecar.system :as fig]
+    [com.stuartsierra.component :as component]))
 
 (defn conform! [spec x]
   (let [rt (s/conform spec x)]
@@ -15,24 +16,23 @@
                (s/explain-data spec x))))
     rt))
 
-(def figwheel-config
-  {:figwheel-options {:server-port 3050
-                      :css-dirs    ["resources/public/css"]}
-   :default-build-ids ["test" "cards"]
-   :all-builds        (figwheel-sidecar.repl/get-project-cljs-builds)})
+(def figwheel (atom nil))
 
 (defn start-figwheel
-  "Start Figwheel on the given builds, or defaults to `default-build-ids` in `figwheel-config`."
-  ([] (let [props (System/getProperties)
-            all-builds (->> figwheel-config :all-builds (mapv :id))]
-        (start-figwheel (keys (select-keys props all-builds)))))
+  "Start Figwheel on the given builds, or defaults to build-ids in `figwheel-config`."
+  ([]
+   (let [figwheel-config (fig/fetch-config)
+         props (System/getProperties)
+         all-builds (->> figwheel-config :data :all-builds (mapv :id))]
+     (start-figwheel (keys (select-keys props all-builds)))))
   ([build-ids]
-   (let [build-ids (or (seq build-ids) (:default-build-ids figwheel-config))]
-     (assert (set/subset? (set build-ids)
-                          (set (map :id (:all-builds figwheel-config))))
-             (str "\nInvalid build ids: " (set/difference (set build-ids)
-                                                          (set (map :id (:all-builds figwheel-config))))
-                  "\nValid ids are: " (set (map :id (:all-builds figwheel-config)))))
+   (let [figwheel-config (fig/fetch-config)
+         default-build-ids (-> figwheel-config :data :build-ids)
+         build-ids (if (empty? build-ids) default-build-ids build-ids)
+         preferred-config (assoc-in figwheel-config [:data :build-ids] build-ids)]
+     (reset! figwheel (component/system-map
+                        :figwheel-system (fig/figwheel-system preferred-config)
+                        :css-watcher (fig/css-watcher {:watch-paths ["resources/public/css"]})))
      (println "STARTING FIGWHEEL ON BUILDS: " build-ids)
-     (ra/start-figwheel! (assoc figwheel-config :build-ids build-ids))
-     (ra/cljs-repl))))
+     (swap! figwheel component/start)
+     (fig/cljs-repl (:figwheel-system @figwheel)))))