+++ src/untangled/client/logging.cljs
@@ -17,16 +17,22 @@
   (str msg ":\n" (with-out-str (cljs.pprint/pprint val))))
 
 (defn debug
-  "Print a debug message which includes a value.
+  "Print a debug message to the Om logger which includes a value.
   Returns the value (like identity) so it can be harmlessly nested in expressions."
   ([value] (glog/fine *logger* (value-message "DEBUG" value)) value)
   ([msg value] (glog/fine *logger* (value-message msg value)) value))
 
-(defn info [& data]
+(defn info
+  "output an INFO level message to the Om logger"
+  [& data]
   (glog/info *logger* (apply str (interpose " " data))))
 
-(defn warn [& data]
+(defn warn
+  "output a WARNING level message to the Om logger"
+  [& data]
   (glog/warning *logger* (apply str (interpose " " data))))
 
-(defn error [& data]
+(defn error
+  "output an ERROR level message to the Om logger"
+  [& data]
   (glog/error *logger* (apply str (interpose " " data))))